ml-service/requirements.txt
fastapi==0.109.0
python-consul==1.1.0
aio-pika==9.4.0
uvicorn==0.27.0
pydantic==2.5.3
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.3.2
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pytest==8.0.0
httpx==0.26.0
pytest-asyncio==0.23.5
